chore: mypage 폴더 생성, router 수정
