+++ src/pages/mypage/MyFootInfo.tsx
@@ -1,6 +1,6 @@
-import Header from '../components/common/Header';
-import myfoot from '../assets/images/foot-loading.gif';
-import Button from '../components/common/Button';
+import Header from '../../components/common/Header';
+import myfoot from '../../assets/images/foot-loading.gif';
+import Button from '../../components/common/Button';
 import { useEffect } from 'react';
 
 const win = window as any;
